+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### 1.55
   * Bugfix:  #439 fix cursor timeouts in chunkstore iterator
+  * Bugfix:  #450 fix error in chunkstore delete when chunk range produces empty df
 
 ### 1.54 (2017-10-18)
   * Bugfix:  #440 Fix read empty MultiIndex+tz Series