+++ CHANGES.md
@@ -4,6 +4,7 @@
   * Bugfix:  #439 fix cursor timeouts in chunkstore iterator
   * Bugfix:  #450 fix error in chunkstore delete when chunk range produces empty df
   * Bugfix:  #442 fix incorrect segment values in multi segment chunks in chunkstore
+  * Feature: #457 enchances fix for #442 via segment_id_repair tool
 
 ### 1.54 (2017-10-18)
   * Bugfix:  #440 Fix read empty MultiIndex+tz Series