+++ CHANGES.md
@@ -6,6 +6,7 @@
   * Bugfix:  #442 fix incorrect segment values in multi segment chunks in chunkstore
   * Feature: #457 enchances fix for #442 via segment_id_repair tool
   * Bugfix:  #385 exceptions during quota statistics no longer kill a write
+  * Feature:  PR#161 TickStore.max_date now returns a datetime in the 'local' timezone
 
 ### 1.54 (2017-10-18)
   * Bugfix:  #440 Fix read empty MultiIndex+tz Series